Fix some errors
앨범 삭제시 삭제 모달
로그 아웃 시 어럴트 띄우기
출석 계산 로직 변경
출석 선생님쪽 css 안먹힌거 확인
요청사항 글 삭제, 수정, 생성시 버튼 효과 첫번째 글로 먹히게 수정
인증 완료 문구 어럴트 처리, 인증 완료 버튼 disabled 처리
비밀번호 변경시 입력 에러 처리:

+++ frontend/src/components/Album/AlbumDetail.vue
@@ -170,21 +170,33 @@ export default {
     },
 
     async deleteAlbum() {
-      let accessToken = sessionStorage.getItem("access-token");
-      let refreshToken = sessionStorage.getItem("refresh-token");
+      this.$fire({
+        html: `<a href="javascript:void(0);"></a><p style="font-size: 0.95rem; font-family: 'NanumSquareRound';">삭제 하시겠습니까?</p>`,
+        type: "question",
+        showCancelButton: true,
+        confirmButtonText: "예",
+        cancelButtonText: "아니오",
+        confirmButtonColor: "#58679A",
+      }).then(async (r) => {
+        if (r.value) {
+          let accessToken = sessionStorage.getItem("access-token");
+          let refreshToken = sessionStorage.getItem("refresh-token");
+    
+          let data = {
+            albumId: this.selectedAlbumId,
+          };
+          await albumApi.deleteAlbum(data, {
+            "access-token": accessToken,
+            "refresh-token": refreshToken,
+          });
+          awss3.deletePhoto(this.photoKeyList, this.downloadUrl);
+          this.photoKeyList = [];
+          this.downloadUrl = "";
+          this.$emit("updateVisible", !this.visible);
+          this.$emit("get-album");
+        }
+      })
 
-      let data = {
-        albumId: this.selectedAlbumId,
-      };
-      await albumApi.deleteAlbum(data, {
-        "access-token": accessToken,
-        "refresh-token": refreshToken,
-      });
-      awss3.deletePhoto(this.photoKeyList, this.downloadUrl);
-      this.photoKeyList = [];
-      this.downloadUrl = "";
-      this.$emit("updateVisible", !this.visible);
-      this.$emit("get-album");
     },
   },
   created() {
